third_party/blink/web_tests/http/tests/websocket/binary-type-expected.txt

CONSOLE WARNING: The provided value 'Blob' is not a valid enum value of type BinaryType.
CONSOLE WARNING: The provided value 'ArrayBuffer' is not a valid enum value of type BinaryType.
CONSOLE WARNING: The provided value '' is not a valid enum value of type BinaryType.
Test WebSocket.binaryType attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ws.binaryType is "blob"
PASS ws.binaryType is "arraybuffer"
PASS ws.binaryType is "blob"
Set invalid values to binaryType. They should be ignored. No exception should be thrown.
PASS ws.binaryType is "blob"
PASS ws.binaryType is "blob"
PASS ws.binaryType is "blob"
PASS successfullyParsed is true

TEST COMPLETE
